requirements.txt
gensim==4.3.2
nltk==3.8.1
underthesea==6.8.0
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1